4-scan_ai-Manus/image_crawler/requirements.txt
# Image Crawler Requirements - Gaara Scan AI v4.3.1

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3

# HTTP Client & Scraping
httpx==0.26.0
requests==2.31.0
beautifulsoup4==4.12.3
selenium==4.17.2
playwright==1.41.0

# Image Processing
pillow==10.2.0
opencv-python-headless==4.9.0.80

# AI/ML for Image Analysis
openai==1.10.0
anthropic==0.8.1

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25

# Redis
redis==5.0.1

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
aiofiles==23.2.1
